Add files via upload
Commented Python program that reads a DHT22 Temperature & Humidity sensor, calculates Temperature in Fahrenheit (from Celsius) as well as Dew Point and Heat Index (in Celsius).  Heat Index method is a full on version of formula used by NOAA and posted on their website.  This code works on a Raspberry Pi Zero WH with RaspiOS 10 (BUSTER).